requirements.txt
vk_api>=11.9.9, <12.0.0
apscheduler>=3.9.1, <4.0
progressbar2>=4.0.0, <5.0.0